requirements.txt
absl-py==0.11.0
appdirs==1.4.4
argcomplete==1.12.1
astor==0.8.1
audioread==2.1.9
backcall==0.2.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
filelock==3.0.12
future==0.18.2
gast==0.2.2
gdown==3.12.2
google-pasta==0.2.0
grpcio==1.33.2
h5py==3.1.0
idna==2.10
inflect==5.0.2
ipython-genutils==0.2.0
jamo==0.4.1
kaldiio==2.17.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
librosa==0.8.0
llvmlite==0.34.0
Markdown==3.3.3
matplotlib==3.3.3
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
numba==0.51.2
numpy==1.21.1
opt-einsum==3.3.0
parallel-wavegan==0.4.8
parso==0.7.0
Pillow==8.0.1
pooch==1.2.0
protobuf==3.14.0
ptyprocess==0.6.0
pycparser==2.20
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.2
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.25.0
resampy==0.2.2
scikit-learn==0.23.2
scipy==1.5.4
SoundFile==0.10.3.post1
tensorboard==1.15.0
tensorboardX==2.1
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.7.0
tqdm==4.53.0
Unidecode==1.1.1
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
xmltodict==0.12.0
yq==2.11.1
transformers==2.10.0
konlpy
torchvision
dominate>=2.3.1
dill
scikit-image
opencv-python>=3.4.3.18
lws
face-alignment
black

